backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.34.13
botocore==1.34.13
cachetools==5.3.0
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect2==0.7.0
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
facebook-sdk==3.1.0
frozenlist==1.3.3
google-auth==2.16.3
google-auth-oauthlib==1.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kombu==5.3.1
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.39
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator
PyJWT==2.6.0
pypdf==3.7.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
razorpay==1.3.0
redis==4.6.0
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.10.0
scipy==1.10.1
sentry-sdk==1.32.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
stripe==5.4.0
svglib==1.5.1
threadpoolctl==3.2.0
tinycss2==1.2.1
twilio==8.4.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.9
yarl==1.9.2
